Update calculator.c
Converting x, y and z variables to `double` data type to have more accuracy in the calculation than the `foat` data type.

+++ Lecture 1 — C/code examples/calculator.c
@@ -6,6 +6,6 @@ int main(void)
   int x = get_int("x: ");
   int y = get_int("y: ");
 
-  float z = (float) x / (float) y;
+  double z = (double) x / (double) y;
   printf("%20f\n", z);
 }
